[WIP] Stark Implementation docs
